fix: resolve 3 critical + 6 warning API consistency issues from round 10 review
Critical fixes:
- R10-1: Rename Prometheus kernel metrics to remove redundant _total suffix
  (firewall_kernel_total_bans_total -> firewall_kernel_bans_total)
  (firewall_kernel_total_unbans_total -> firewall_kernel_unbans_total)
- R10-3: Split systemd ReadWritePaths to follow least-privilege principle
  (/etc/firewall moved to ReadOnlyPaths, /var/lib/firewall kept in ReadWritePaths)

Warning fixes:
- R10-4: Refactor top-level YAML config parsing to reuse apply_defaults_config
  (eliminated ~170 lines of duplicated parsing logic)
- R10-5: Add has_error parameter to apply_jail_string_config for error tracking
- R10-6: Add detailed error messages to config_validate for each failure case
- R10-7: Change ip_to_str buffer length parameter from int to size_t (kernel standard)
- R10-8: Fix config_write to check parameter name before parsing numeric value
- R10-9: Remove implicit permanent_ban_enabled=1 when setting permanent_db_path

+++ src/kernel-module/firewall.h
@@ -292,7 +292,7 @@ extern u32 fw_hash_seed;
  * @buf: 输出缓冲区 (至少 INET6_STR_LEN 字节)
  * @len: 缓冲区长度
  */
-static inline void ip_to_str(u8 af, const void *ip, char *buf, int len) {
+static inline void ip_to_str(u8 af, const void *ip, char *buf, size_t len) {
   if (af == FW_AF_INET6) {
     const struct in6_addr *addr = ip;
     if (len < INET6_STR_LEN) {
